refactor: DB 연결 클래스 위치 변경
